// src/app/shared/generated/model/colorScheme.ts
/**
 * onecx-user-profile-bff
 * Backend-For-Frontend (BFF) service for onecx-user-profile. This API provides endpoints to manage User Profiles, Account Settings and Avatars
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export enum ColorScheme {
    Auto = 'AUTO',
    Light = 'LIGHT',
    Dark = 'DARK'
}