// src/app/shared/generated/model/updateUserSettings.ts
/**
 * onecx-user-profile-bff
 * OneCx user profile Bff
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ColorScheme } from './colorScheme';
import { MenuMode } from './menuMode';


export interface UpdateUserSettings { 
    modificationCount?: number;
    hideMyProfile?: boolean;
    locale?: string;
    timezone?: string;
    menuMode?: MenuMode;
    colorScheme?: ColorScheme;
}